all UI

+++ academity_app/lib/services/auth_services.dart
@@ -13,8 +13,8 @@ class AuthServices {
   Future<bool> login(String email, String password) async {
     // here we are not using AcademityApi.get because we don't have a token
     // and /api/login does not need one
-    final response = await http
-        .post(Uri.http(Env.academityHost, '/api/login'), body: {
+    final response =
+        await http.post(Uri.http(Env.academityHost, '/api/login'), body: {
       'email': email,
       'password': password,
     });
@@ -31,9 +31,37 @@ class AuthServices {
     return false;
   }
 
+  Future<bool> registerUser(Map<String, dynamic> data) async {
+    final response = await http.post(
+      Uri.http(Env.academityHost, '/api/register'),
+      body: data,
+    );
+
+    if (response.statusCode == 200) {
+      final responseData = json.decode(response.body);
+      if (responseData['status'] == 'Register successful') {
+        const secureStorage = FlutterSecureStorage();
+        final apiToken = responseData['new_token'];
+        await secureStorage.write(key: 'api_token', value: apiToken);
+        return true;
+      }
+    }
+    return false;
+  }
+
   Future<void> logout() async {
-    // Handle logout logic, possibly including clearing shared preferences
+    final response = await AcademityApi.post('logout');
     const secureStorage = FlutterSecureStorage();
-    await secureStorage.delete(key: 'api_token');
+
+    if (response.statusCode == 200 || response.statusCode == 204) {
+      // Successfully logged out, delete the token
+      await secureStorage.delete(key: 'api_token');
+    } else {
+      // Log the response for debugging
+      print(
+          'Logout failed with status code: ${response.statusCode} and body: ${response.body}');
+      throw Exception(
+          'Failed to logout, Server responded with status code: ${response.statusCode}');
+    }
   }
 }
